# 07/doc/openapi.json
{"openapi":"3.0.1","info":{"title":"Conditional GET tours REST service","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/tour/{id}":{"get":{"tags":["tour-controller"],"operationId":"getTour","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tour"}}}}}},"delete":{"tags":["tour-controller"],"operationId":"deleteTour","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/tour":{"get":{"tags":["tour-controller"],"operationId":"getTours","parameters":[{"name":"eTagValidation","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelTour"}}}}}},"post":{"tags":["tour-controller"],"operationId":"createTour","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tour"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Link":{"type":"object","properties":{"rel":{"type":"string"},"href":{"type":"string"},"hreflang":{"type":"string"},"media":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"}}},"Tour":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"customers":{"type":"array","items":{"type":"string"}},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"CollectionModelTour":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}},"content":{"type":"array","items":{"$ref":"#/components/schemas/Tour"}}}}}}}